LayoutTests/accessibility/gtk/text-for-range-combo-box-expected.txt


This tests the ability to get element text for a specified range.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS combobox.stringForRange(0, -1) is ''
PASS popup.stringForRange(0, -1) is ''
PASS option.stringForRange(0, -1) is 'An option in a combo box'
PASS combobox.stringForRange(0, -1) is ''
PASS popup.stringForRange(0, -1) is ''
PASS option.stringForRange(0, -1) is 'foo'
PASS option.stringForRange(0, 2) is 'fo'
PASS option.stringForRange(0, -1) is 'bar'
PASS option.stringForRange(1, 3) is 'ar'
PASS successfullyParsed is true

TEST COMPLETE
